Juanmagw

+++ app/Http/Controllers/UserController.php
@@ -176,7 +176,7 @@ class UserController extends Controller
      *         description="Shows all the information about of a user based that matches an id"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
@@ -218,7 +218,7 @@ class UserController extends Controller
      *         description="Shows all the information about of a user based that matches an dni"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
@@ -266,7 +266,7 @@ class UserController extends Controller
      *         description="Shows all the information about of a user based that matches an mail"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
@@ -314,7 +314,7 @@ class UserController extends Controller
      *         description="Shows all the information about of a user based that matches an course_year"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
@@ -332,7 +332,7 @@ class UserController extends Controller
      *         description="List all the students of the database"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
@@ -345,7 +345,7 @@ class UserController extends Controller
      *         description="List all the students of the database"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
@@ -435,7 +435,7 @@ class UserController extends Controller
      *         description="Shows all the information about of a user based that matches an cycle"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
@@ -465,6 +465,48 @@ class UserController extends Controller
         ], 401);
     }
 
+    public function addUser(Request $request)
+    {
+        if ($request->user('api')->rol == '0' || $request->user('api')->rol == '1') {
+            $user = new User();
+            $user->dni = $request->dni;
+            if ($request->type == 'professor') {
+                $user->rol = 1;
+            } else if ($request->type == 'student') {
+                $user->rol = 2;
+            } else {
+                return response()->json([
+                    'status' => -1,
+                    'message' => 'No type found',
+                ], 400);
+            }
+            $user->course_year = $request->course_year;
+            $user->cycle = $request->cycle;
+            $user->name = $request->name;
+            $user->surname = $request->surname;
+            $user->email = $request->email;
+            $user->password = Hash::make('root');
+            $user->others = $request->others;
+
+            if ($user->save()) {
+                return response()->json([
+                    'status' => 1,
+                    'message' => 'User added',
+                ], 200);
+            } else if (!$user->save()) {
+                return response()->json([
+                    'status' => -1,
+                    'message' => 'User not added',
+                ], 400);
+            }
+        }
+
+        return response()->json([
+            'status' => -1,
+            'message' => 'Unauthorized',
+        ], 401);
+    }
+
     /**
      * Adds a new student to the database.
      *
@@ -477,7 +519,7 @@ class UserController extends Controller
      *         description="Adds a new student to the database"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
@@ -509,7 +551,7 @@ class UserController extends Controller
      *         description="Adds a new professor to the database"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
@@ -542,7 +584,7 @@ class UserController extends Controller
      *         description="Adds all students to the database from a json file"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
@@ -563,7 +605,7 @@ class UserController extends Controller
      *         description="Adds all professor to the database from a json file"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
@@ -590,70 +632,45 @@ class UserController extends Controller
      *         description="Update an user using their id as reference"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
      */
     public function editUser(Request $request)
     {
-        $user = User::findOrFail($request->id);
-        $user->dni = $request->dni;
-        $user->course_year = $request->course_year;
-        $user->cycle = $request->cycle;
-        $user->name = $request->name;
-        $user->surname = $request->surname;
-        $user->email = $request->email;
-        if (strlen($request->password) > 8) {
-            $user->password = Hash::make($request->password);
-            return response()->json([
-                'status' => 1,
-                'message' => 'Password accepted.',
-            ], 200);
-        } elseif (strlen($request->password) >= 1) {
-            return response()->json([
-                'status' => -1,
-                'message' => 'Password must have at least 8 characters.',
-            ], 400);
+        if ($request->user('api')->rol == '0' || $request->user('api')->rol == '1') {
+            $user = User::findOrFail($request->id);
+            $user->dni = $request->dni;
+            $user->course_year = $request->course_year;
+            $user->cycle = $request->cycle;
+            $user->name = $request->name;
+            $user->surname = $request->surname;
+            $user->email = $request->email;
+            if (strlen($request->password) > 8) {
+                $user->password = Hash::make($request->password);
+                return response()->json([
+                    'status' => 1,
+                    'message' => 'Password accepted.',
+                ], 200);
+            } elseif (strlen($request->password) >= 1) {
+                return response()->json([
+                    'status' => -1,
+                    'message' => 'Password must have at least 8 characters.',
+                ], 400);
+            }
+            $user->others = $request->others;
+
+            $user->save();
+
+            return $user;
         }
-        $user->others = $request->others;
 
-        $user->save();
-
-        return $user;
+        return response()->json([
+            'status' => -1,
+            'message' => 'Unauthorized',
+        ], 401);
     }
-
-    public function editStudent(Request $request)
-    {
-
-        $user = User::find($request->id);
-        $user->dni = $request->dni;
-        $user->course_year = $request->course_year;
-        $user->cycle = $request->cycle;
-        $user->name = $request->name;
-        $user->surname = $request->surname;
-        $user->email = $request->email;
-        if (strlen($request->password) > 8) {
-            $user->password = Hash::make($request->password);
-            return response()->json([
-                'status' => 1,
-                'message' => 'Password accepted.',
-            ], 200);
-        } elseif (strlen($request->password) >= 1) {
-            return response()->json([
-                'status' => -1,
-                'message' => 'Password must have at least 8 characters.',
-            ], 400);
-        }
-        $user->others = $request->others;
-
-        $user->save();
-
-        return $user;
-
-    }
-
-
 
     /**
      * Remove an user based on their id
@@ -673,26 +690,33 @@ class UserController extends Controller
      *         description="Remove an user using their id as reference"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
      */
     public function deleteUser(Request $request)
     {
-        $id = $request->id;
-        $users = User::destroy($id);
-        if ($users) {
+        if ($request->user('api')->rol == '0' || $request->user('api')->rol == '1') {
+            $id = $request->id;
+            $users = User::destroy($id);
+            if ($users) {
+                return response()->json([
+                    'status' => 1,
+                    'message' => 'Delete user by ID ' . $id,
+                ], 200);
+            }
+
             return response()->json([
-                'status' => 1,
-                'message' => 'Delete user by ID ' . $id,
-            ], 200);
+                'status' => -1,
+                'message' => 'No User Found',
+            ], 400);
         }
 
         return response()->json([
             'status' => -1,
-            'message' => 'No User Found',
-        ], 400);
+            'message' => 'Unauthorized',
+        ], 401);
     }
 
     /**
@@ -713,7 +737,7 @@ class UserController extends Controller
      *         description="Remove an user using their rol as reference"
      *     ),
      *     @OA\Response(
-     *         response=404,
+     *         response=400,
      *         description="An error has ocurred."
      *     )
      * )
